Updated all tests to use the new Color struct
