+++ embedded-graphics/src/fonts/font12x16.rs
@@ -4,6 +4,7 @@ use super::super::drawable::*;
 use super::super::transform::*;
 use super::Font;
 use coord::{Coord, ToUnsigned};
+use color::Color;
 
 const FONT_IMAGE: &[u8] = include_bytes!("../../data/font12x16_1bpp.raw");
 const CHAR_HEIGHT: u32 = 16;
@@ -149,8 +150,9 @@ impl<'a, C> Transform for Font12x16<'a, C> where C: Clone + Copy + PartialEq {
     /// # use embedded_graphics::fonts::{ Font, Font12x16 };
     /// # use embedded_graphics::transform::Transform;
     /// # use embedded_graphics::coord::Coord;
+    /// # use embedded_graphics::color::Color;
     ///
-    /// let text = Font12x16::render_str("Hello world", 1);
+    /// let text = Font12x16::render_str("Hello world", Color::new(1));
     /// let moved = text.translate(Coord::new(25, 30));
     ///
     /// assert_eq!(text.pos, Coord::new(0, 0));
@@ -169,8 +171,9 @@ impl<'a, C> Transform for Font12x16<'a, C> where C: Clone + Copy + PartialEq {
     /// # use embedded_graphics::fonts::{ Font, Font12x16 };
     /// # use embedded_graphics::transform::Transform;
     /// # use embedded_graphics::coord::Coord;
+    /// # use embedded_graphics::color::Color;
     ///
-    /// let mut text = Font12x16::render_str("Hello world", 1);
+    /// let mut text = Font12x16::render_str("Hello world", Color::new(1));
     /// text.translate_mut(Coord::new(25, 30));
     ///
     /// assert_eq!(text.pos, Coord::new(25, 30));
