+++ embedded-graphics/src/fonts/mod.rs
@@ -10,7 +10,7 @@ pub use self::font6x12::Font6x12;
 pub use self::font8x16::Font8x16;
 pub use self::font6x8::Font6x8;
 
-use drawable::Color;
+use color::Color;
 
 /// Common methods for all fonts
 pub trait Font<'a> {
@@ -23,12 +23,15 @@ pub trait Font<'a> {
     /// # use embedded_graphics::fonts::{Font, Font6x8};
     /// # use embedded_graphics::transform::Transform;
     /// # use embedded_graphics::drawable::Pixel;
+    /// # use embedded_graphics::color::Color;
+    /// #
     /// #
     /// # struct Display {}
     /// # impl Display {
-    /// #     pub fn draw<T>(&self, item_pixels: T) -> Result<(), ()>
+    /// #     pub fn draw<T, C>(&self, item_pixels: T) -> Result<(), ()>
     /// #     where
-    /// #         T: Iterator<Item = Pixel>,
+    /// #         T: Iterator<Item = Pixel<C>>,
+    /// #         C: Clone + Copy + PartialEq,
     /// #     {
     /// #         Ok(())
     /// #     }
@@ -37,7 +40,7 @@ pub trait Font<'a> {
     /// fn main() {
     ///     let disp = Display {};
     ///     // Render a string with a 8bit color
-    ///     let text = Font6x8::render_str("Hello world", 1);  
+    ///     let text = Font6x8::render_str("Hello world", Color::new(1));  
     ///
     ///     disp.draw(text.into_iter());
     /// }
