+++ embedded-graphics/src/image/image1bpp.rs
@@ -12,6 +12,7 @@ use super::super::drawable::*;
 use super::super::transform::*;
 use super::Image;
 use coord::{Coord, ToUnsigned};
+use color::Color;
 
 /// 1 bit per pixel image
 #[derive(Debug)]
